+++ R/simulateCamtrapData.R
@@ -67,7 +67,7 @@
 #'                                 setupCol = "Setup_date",
 #'                                 retrievalCol = "Retrieval_date")
 #' 
-#' camtrapR:::camopPlot(camop_simple)
+#' plot(camop_simple)
 #' 
 #' 
 #' 
@@ -100,8 +100,10 @@
 #'                                  allCamsOn = FALSE,
 #'                                  camerasIndependent = FALSE,   # this is made up, not simulated
 #'                                  dateFormat = "ymd HMS")
+#'                                  
+#' summary(camop_complex)
 #' 
-#' camtrapR:::camopPlot(camop_complex)
+#' plot(camop_complex)
 
 simulateCamtrapData <- function(nStations = 10,
                                 camerasPerStation = 1,
@@ -279,6 +281,14 @@ simulateCamtrapData <- function(nStations = 10,
   
   recordTable <- recordTable[order(recordTable$Station, recordTable$DateTimeOriginal), ]
   rownames(recordTable) <- NULL
+  
+  # declare specific classes and store attributes
+  class(camtraps) <- unique(c("cams", class(camtraps)))
+  attr(camtraps, "stationCol") <- "Station"
+  
+  class(recordTable) <- unique(c("records", class(recordTable)))
+  attr(recordTable, "stationCol") <- "Station"
+  attr(recordTable, "speciesCol") <- "Species"
   
   return(list(camtraps = camtraps, recordTable = recordTable))
 }